complete SingleColorPicker.js
